+++ src/main/java/me/gabcytn/srsly/Repository/SolutionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SolutionRepository extends ListCrudRepository<Solution, Integer> {
+public interface SolutionRepository extends ListCrudRepository<Solution, Long> {
   Boolean existsByProblemAndUser(Problem problem, User user);
 
   Integer countByProblemAndUser(Problem problem, User user);
